Folder rename. PropTypes
