+++ src/components/Share/SharePanel.js
